update adpaters to use Ember.$.ajax for XHRs
due to using DataAdapterMixin, which adds auth headers in the
ajax.beforeSend hook. if the request is called using fetch, that hook
won't run.

after updating adapters to use $.ajax, it turned out that jQuery's
resolution wasn't a true promise, which caused acceptance tests to fail.

This patch adds a `wrap-ajax` module to manually resolve the $.ajax
calls.

We could update fetch calls to use ESA's `authorize` call to add the
headers manually. Maybe we should. this works for now.