update user model with cognito attrs
the auth service will return User models in JSONAPI format, but for now
only accepts plain JSON objects. the serializer will strip the outgoing
JSONAPI formatting and only send along the `attributes` key.

this updates the serializer to transform the attribute key to underscore
from camelcase when the keys are going out to the network.