only send changed attrs on update requests
there's a standards doc that proposes only sending the changed data in
PUT & PATCH requests, which works for us, so this implements that
strategy for the user model. this was to get around 409 responses from
the auth server erroneously recognizing an incoming `preferred_username`
field as a duplicate, when it's just the value owned by the incoming
user trying to update other attrs.

see here for more: https://tools.ietf.org/html/rfc7386