use an observer to validate emailConfirmation
I don't love it, but emailConfirmation is not a field on the user model,
so there's no way to trigger the valiation on it unless we do it
manually. Or else we could refactor the templates to call a general
purpose "validate" method on input for all the inputs.

barring that, this is a way to run the validation on the confirm email
field whenever the email field value changes.