adds torii provider and authenticator
torii is used to authenticate against third-party auth providers like
facebook. it requires torii is added as a dependency and that the
FacebookConnect provider is implemented with a bit of boilerplate for
ember-simple-auth's uses.

the other bit this changes is overriding the ember simple auth Torii
authenticator's authenticate method to return an authenticated data
object that conforms to how the rest of the app needs it.

one thing this does is include a value for the `provider` key, which
will be passed to the auth service at the `X-Provider` header. the
service uses that value to determine how to validate the provided access
token which is added at the Authorization header, as normal.