sign up with facebook action
this is currently on the signup route, but should probably be moved to
the account sign up componente per @walsh9's prerogative. Here's the
logic:

* use the torii provider directly to go through facebook auth, instead
of using the session#authenticate method. this prevents the global
sessionAuthenticated event from firing, which has undesired consequences.
* once authenticated, make a call against the FB api using the FB sdk in
order to gather the required fields for a new user
* create the user, save the user
* now use the session's authenticate method so that it triggers the
global sessionAuthenticated event, which will load up the current user

*NOTE*: if we end up refactoring this to so that the current user is
**not** loaded during the sessionAuthenticated, then the logic can
change so that first we authenticate using the session, and in the
callback create the user. if this occurs, then the work in 1451399a
can be taken out, since the headers added in that commit will
automatically get added since the session will be authenticated.