enforce that start actions always send 0 as the current position
we don't fire `start` until hifi has reported that the audio is playing.
while this is reliable, it's not immediate. by the time
`sendListenAction` is triggered, the real-time position of the audio is
not zero; it's usually around 4s. this patch enforces that all start
actions report a position as 0, even if that's not totally accurate.

this might cause inconsistent current_position values when looking at
resumes followed by a pause. the current_position for a pause action will
properly report, but the resume will suffer the same delay as a start.
I would expect that resume actions will report their current_position as
~4s later than the current_position reported for its associated pause
action.